Incremental updates.
Fixing the location which had two address fields with different names.
Cleaning up some of the HTML in the form pages so that they don't have a wideness problem for the container with the submit and back buttons.
Adding the description information to the project_detail page.
Adding a new Project copy template so that the user can see what project it is that they are copying.